arrange structure
