--- src/assets/scss/Home.scss
@@ -1,104 +1,0 @@
-.header {
-  display: flex;
-  position: fixed;
-  align-items: center;
-  justify-content: center;
-  height: 5rem;
-  width: 100%;
-  top: 0;
-  z-index: 1;
-  background-color: rgba(0, 0, 0, 0);
-
-  .button_login {
-    position: absolute;
-    float: right;
-    right: 1rem;
-  }
-  .header_logo {
-    width: 5rem;
-    position: absolute;
-    float: left;
-    left: 4rem;
-    top: 0.5rem;
-  }
-}
-.button_text {
-  text-decoration: none;
-  color: white;
-  padding: 0 1rem 0 0;
-}
-.logo {
-  position: relative;
-  width: 30rem;
-  align-items: center;
-  left: 29rem;
-  top: 13rem;
-}
-
-.logo_desc {
-  color: #ffffff;
-  font-size: large;
-  font-family: "GmarketSansTTF Medium";
-  width: 13rem;
-  text-align: center;
-  position: relative;
-  left: 43rem;
-  top: 10.5rem;
-}
-.scroll_next {
-  color: #ffffff;
-  width: auto;
-  position: absolute;
-  bottom: 1rem;
-  left: 45%;
-  opacity: 0.6;
-}
-
-.section2_desc {
-  width: 60rem;
-  position: relative;
-  left: 14rem;
-  top: 13rem;
-  text-align: center;
-  font-family: "GmarketSansTTF Medium";
-  font-weight: bold;
-  font-size: xx-large;
-  color: #ffffff;
-
-  .section2_desc_bold {
-    font-size: 3.5rem;
-    position: relative;
-    bottom: 1.5rem;
-  }
-}
-
-.section2_button_room {
-  position: relative;
-  left: 35.5rem;
-  top: 14.5rem;
-  display: flex;
-  width: 17rem;
-  justify-content: space-between;
-}
-.section3_photo_desc {
-  width: 34rem;
-  position: relative;
-  left: 40rem;
-  top: 14rem;
-  font-family: "GmarketSansTTF Medium";
-  color: #ffffff;
-
-  .section3_photo_desc_title {
-    font-size: xx-large;
-    font-weight: bold;
-  }
-}
-
-.section3_button_room {
-  position: relative;
-  left: 44rem;
-  top: 15rem;
-  display: flex;
-  width: 17rem;
-  justify-content: space-between;
-}
