+++ src/header/scss/Navbar.scss
@@ -1,0 +1,107 @@
+.navbar {
+  height: 80px;
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  position: absolute;
+  float: left;
+  left: -0.5rem;
+}
+
+.menu-bars {
+  margin-left: 2rem;
+  font-size: 2rem;
+  background-color: none;
+}
+.menu-bars-x {
+  font-size: 2rem;
+  background-color: none;
+}
+
+.nav-menu {
+  background-color: #ff8080;
+  width: 250px;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  left: -100%;
+  transition: 850ms;
+}
+
+.nav-menu.active {
+  left: 0;
+  transition: 350ms;
+  z-index: 1;
+}
+
+.nav-text {
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  padding: 8px 0px 8px;
+  list-style: none;
+  height: 60px;
+}
+
+.nav-text a {
+  text-decoration: none;
+  color: white;
+  font-size: 18px;
+  width: 95%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  padding: 0 8px;
+  border-radius: 4px;
+}
+
+.nav-text a:hover {
+  background-color: #ff4d4d;
+}
+
+.nav-menu-items {
+  width: 100%;
+}
+
+.navbar-toggle {
+  width: 100%;
+  height: 80px;
+  display: flex;
+  justify-content: start;
+  align-items: center;
+}
+.nav_title {
+  margin-left: 16px;
+}
+
+.header {
+  display: flex;
+  position: fixed;
+  align-items: center;
+  justify-content: center;
+  height: 5rem;
+  width: 100%;
+  top: 0;
+  z-index: 1;
+  background-color: rgba(0, 0, 0, 0);
+
+  .button_login {
+    position: absolute;
+    float: right;
+    right: 1rem;
+  }
+  .header_logo {
+    width: 5rem;
+    position: absolute;
+    float: left;
+    left: 4rem;
+    top: 0.5rem;
+  }
+}
+.button_text {
+  text-decoration: none;
+  color: white;
+  text-align: center;
+}
