+++ src/roomList/tileData.js
@@ -1,0 +1,69 @@
+const tileData = [
+    {
+      img: "./pomodoro.png",
+      title: '경주의 공부방',
+      author: '경쥬',
+    },
+    {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+      {
+        img: "./pomodoro.png",
+        title: '경주의 공부방',
+        author: '경쥬',
+      },
+  ];
+
+  export default tileData;
