--- src/view/Navbar/SidebarData.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import * as AiIcons from "react-icons/ai";
-
-export const SidebarData = [
-    {
-        title: 'Home',
-        path: '/',
-        icon: <AiIcons.AiFillHome />,
-        cName: 'nav-text'
-    },
-    {
-        title: 'About',
-        path: '/About',
-        icon: <AiIcons.AiFillBook />,
-        cName: 'nav-text'
-    }
-]
