--- src/view/tileData.js
@@ -1,69 +1,0 @@
-const tileData = [
-    {
-      img: "./pomodoro.png",
-      title: '경주의 공부방',
-      author: '경쥬',
-    },
-    {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-      {
-        img: "./pomodoro.png",
-        title: '경주의 공부방',
-        author: '경쥬',
-      },
-  ];
-
-  export default tileData;
